refactor: update images
