+++ src/components/ResourceBrowser/ResourceList/ConnectingToClusterState.tsx
@@ -18,7 +18,7 @@ import React, { useEffect, useState, useRef } from 'react'
 import { useHistory, useParams } from 'react-router-dom'
 import { TAKING_LONGER_TO_CONNECT, TRYING_TO_CONNECT, CONNECTION_TIMEOUT_TIME } from '../Constants'
 import { ConnectingToClusterStateProps, URLParams } from '../Types'
-import CouldNotConnectImg from '../../../assets/img/app-not-deployed.png'
+import CouldNotConnectImg from '../../../assets/img/app-not-deployed.svg'
 import { StyledProgressBar } from '../../common/formFields/Widgets/Widgets'
 import { URLS } from '../../../config'
